requirements.txt
appdirs==1.4.4
booleanOperations==0.9.0
Brotli==1.1.0
cffsubr==0.3.0
defcon==0.10.3
fontFeatures==1.8.0
fontMath==0.9.4
fonttools==4.54.1
fs==2.4.16
lxml==5.3.0
pyclipper==1.3.0.post5
six==1.16.0
ufo2ft==3.3.1
ufo_extractor==0.8.0
unicodedata2==15.1.0
zopfli==0.2.3
